stt/requirements.txt
fastapi
uvicorn
pydantic
transformers
torchaudio
datasets
soundfile
sentencepiece
websockets
wsproto
pika
pybase64
ffmpeg
ffmpeg-python